+++ src/main/kotlin/com/github/ajalt/colormath/Ansi16.kt
@@ -5,7 +5,7 @@ import kotlin.math.roundToInt
 /**
  * An ANSI-16 color code
  */
-data class Ansi16(val code: Int) : ConvertibleColor {
+data class Ansi16(val code: Int) : Color {
     init {
         require(code in 30..37 || code in 40..47 ||
                 code in 90..97 || code in 100..107) {